requirements.txt
Django>=4.2.0
openai>=1.0.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
gunicorn>=21.2.0
django-environ>=0.11.2
